Move preferences migration code to separate file
