+++ app/src/main/java/com/yacgroup/yacguide/LaunchActivity.kt
@@ -22,12 +22,12 @@ import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
 import android.os.CountDownTimer
-import android.util.Log
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import com.yacgroup.yacguide.database.DatabaseWrapper
 import com.yacgroup.yacguide.extensions.getPackageInfoCompat
 import com.yacgroup.yacguide.extensions.versionCodeCompat
+import com.yacgroup.yacguide.migration.PreferencesMigration
 import com.yacgroup.yacguide.network.CountryAndRegionParser
 import com.yacgroup.yacguide.utils.IntentConstants
 
@@ -53,12 +53,13 @@ class LaunchActivity : AppCompatActivity() {
 
         _db = DatabaseWrapper(this)
         _customSettings = getSharedPreferences(getString(R.string.preferences_filename), Context.MODE_PRIVATE)
-        _checkForSettingsMigration()
+        _checkForPreferencesMigration()
 
         _updateHandler = UpdateHandler(this, CountryAndRegionParser(_db))
         _updateHandler.update(
             onUpdateFinished = { findViewById<ImageView>(R.id.iconImageView).setImageResource(R.drawable.ic_start_done) },
-            isSilent = true)
+            isSilent = true
+        )
         Timer().start()
     }
 
@@ -86,7 +87,8 @@ class LaunchActivity : AppCompatActivity() {
      */
     private fun _checkForVersionUpdate(): Boolean {
         val savedVerCode = _customSettings.getInt(
-            getString(R.string.pref_key_version_code), 0)
+            getString(R.string.pref_key_version_code), 0
+        )
         val curVerCode = try {
             packageManager.getPackageInfoCompat(packageName, 0).versionCodeCompat
         } catch (e: Exception) {
@@ -103,75 +105,15 @@ class LaunchActivity : AppCompatActivity() {
     }
 
     /**
-     * Check whether the settings need to be migrated or not.
+     * Check whether the preferences need to be migrated or not.
      */
-    private fun _checkForSettingsMigration() {
-        if (_customSettings.getInt(getString(R.string.pref_key_pref_version), 1) == 1) {
-            _migrateSettings()
-        }
-    }
-
-    /**
-     * Migrate settings, if necessary, to fix issue https://github.com/YacGroup/yacguide/issues/453.
-     */
-    private fun _migrateSettings() {
-        val settingsResourceMap = mapOf(
-            R.string.order_tourbook_chronologically to
-                    Pair(R.string.pref_key_order_tourbook_chronologically, R.bool.pref_default_order_tourbook_chronologically),
-            R.string.only_official_summits to
-                    Pair(R.string.pref_key_only_official_summits, R.bool.pref_default_only_official_summits),
-            R.string.only_official_routes to
-                    Pair(R.string.pref_key_only_official_routes, R.bool.pref_default_only_official_routes),
-            R.string.count_summits to
-                Pair(R.string.pref_key_count_summits, R.bool.pref_default_count_summits),
-            R.string.count_massifs to
-                Pair(R.string.pref_key_count_massifs, R.bool.pref_default_count_massifs),
-            R.string.count_boulders to
-                    Pair(R.string.pref_key_count_boulders, R.bool.pref_default_count_boulders),
-            R.string.count_caves to
-                    Pair(R.string.pref_key_count_caves, R.bool.pref_default_count_caves),
-            R.string.count_unofficial_rocks to
-                    Pair(R.string.pref_key_count_unofficial_rocks, R.bool.pref_default_count_unofficial_rocks),
-            R.string.count_prohibited_rocks to
-                    Pair(R.string.pref_key_count_prohibited_rocks, R.bool.pref_default_count_prohibited_rocks),
-            R.string.count_collapsed_rocks to
-                    Pair(R.string.pref_key_count_collapsed_rocks, R.bool.pref_default_count_collapsed_rocks),
-            R.string.count_only_leads to
-                    Pair(R.string.pref_key_count_only_leads, R.bool.pref_default_count_only_leads),
-            R.string.colorize_tourbook_entries to
-                    Pair(R.string.pref_key_colorize_tourbook_entries, R.bool.pref_default_colorize_tourbook_entries),
-            R.string.lead to
-                    Pair(R.string.pref_key_color_lead, R.color.pref_default_color_lead),
-            R.string.follow to
-                    Pair(R.string.pref_key_color_follow, R.color.pref_default_color_follow)
+    private fun _checkForPreferencesMigration() {
+        val preferencesMigration = PreferencesMigration(
+            context = this,
+            customSettings = _customSettings
         )
-        _customSettings.let {
-            it.edit().apply {
-                for ((oldResource, newResourcePair) in settingsResourceMap) {
-                    val oldKey = resources.getString(oldResource)
-                    val newKey = resources.getString(newResourcePair.first)
-                    Log.d("[PREF]", "Checking for deprecated key '$oldKey'.")
-                    if (it.contains(oldKey)) {
-                        Log.d("[PREF]", "Migrating key '$oldKey' to '$newKey'.")
-                        if (oldResource in listOf(R.string.lead, R.string.follow)) {
-                            this.putInt(
-                                newKey,
-                                it.getInt(oldKey, this@LaunchActivity.getColor(newResourcePair.second))
-                            )
-                        } else {
-                            this.putBoolean(
-                                newKey,
-                                it.getBoolean(oldKey, resources.getBoolean(newResourcePair.second))
-                            )
-                        }
-                        this.remove(oldKey)
-                    }
-                }
-                this.putInt(
-                    resources.getString(R.string.pref_key_pref_version),
-                    resources.getInteger(R.integer.pref_version)
-                )
-            }.apply()
+        if (_customSettings.getInt(getString(R.string.pref_key_pref_version), 1) == 1) {
+            preferencesMigration.migrateFromVersion1To2()
         }
     }
 }
