+++ app/src/main/java/com/yacgroup/yacguide/RockActivity.kt
@@ -80,7 +80,7 @@ class RockActivity : TableActivityWithOptionsMenu() {
         _searchBarHandler = SearchBarHandler(
             searchBarLayout = findViewById(R.id.searchBarLayout),
             searchHintResource = R.string.rock_search,
-            _checkBoxTitle = getString(R.string.only_official_summits),
+            checkBoxTitle = getString(R.string.only_official_summits),
             checkBoxDefaultValue = resources.getBoolean(R.bool.pref_default_only_official_summits),
             _settings = customSettings,
             _settingsKey = getString(R.string.pref_key_only_official_summits),
